+++ src/Engine/Input/Mouse.h
@@ -18,7 +18,7 @@ class Mouse : public InputDevice {
     void resetWheelDelta();
 
     void onPosChange(Vector2 pos);
-    void onRawMove(int xDelta, int yDelta, bool absolute = false, bool virtualDesktop = false);
+    void onRawMove(float xDelta, float yDelta, bool absolute = false, bool virtualDesktop = false);
 
     void onWheelVertical(int delta);
     void onWheelHorizontal(int delta);
